visualizer/requirements.txt
h5py>=3.0.0
numpy>=1.22
matplotlib>=3.0
Pillow>=8.0
PyYAML>=5.3
